Implement joins between globs, expressions, and paths.
This change introduces the `join` function (and `Join` trait). This
function allows globs, expressions, and paths to be joined in a similar
manner to paths via `Path::join`.

To support this change, `Tokenized` now generates its expression when it
is not parsed directly (e.g., when it is the result of a join).
Generated expressions are accurate but lossy, because they may differ
from the given expression. In particular, flags and their interactions
with literals are verbose.