[broken] Implement iterator adaptors for right-most traversal.
This dicey change uses complex adaptors to inspect positional data in
`WalkEntry`s to perform generic (over tokens or components) right-most
traversals of a path in a token tree. This is not quite a degenerate
linear traversal, as it visits all branches of alternatives. This allows
the right-most components of a token tree to be examined, which is
exactly what is needed to query exhaustiveness (there are currently some
false negatives).