Decompose component construction.
This change refactors code that batches tokens into groups as well as
component construction. A macro is used for grouping tokens in order to
preserve the difficult-to-name iterator type used by callers. Components
can be constructed via a flat (equal-depth) naked iterator or via a
tree-traversing iterator using `token::components` and
`Walk::components`, respectively. This is heavily abstracted to share
code.

THIS CHANGE REQUIRES PATCHES TO `itertools` version `0.10.0`!